api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
gunicorn==22.0.0
pydantic==2.7.4
python-dotenv==1.0.1
reportlab==4.2.0
qrcode==7.4.2
Pillow==10.3.0
requests==2.32.3
google-cloud-tasks==2.16.4
google-cloud-logging==3.10.0
redis==5.0.7
google-auth==2.32.0
google-cloud-storage==2.17.0
